requirements.txt
python-telegram-bot==13.15
openai==1.3.5